created UserController and Method Delete in CommentController

+++ src/main/java/br/com/framework/api/frameworkapi/controller/CommentController.java
@@ -9,6 +9,7 @@ import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -22,8 +23,10 @@ import org.springframework.web.util.UriComponentsBuilder;
 import br.com.framework.api.frameworkapi.dto.CommentDto;
 import br.com.framework.api.frameworkapi.form.CommentsPostForm;
 import br.com.framework.api.frameworkapi.model.Comment;
+import br.com.framework.api.frameworkapi.model.User;
 import br.com.framework.api.frameworkapi.repository.CommentRepository;
 import br.com.framework.api.frameworkapi.repository.PostRepository;
+import br.com.framework.api.frameworkapi.repository.UserRepository;
 
 @RestController
 @CrossOrigin(origins = "http://localhost:4200", maxAge = 3600)
@@ -35,34 +38,54 @@ public class CommentController {
 
 	@Autowired
 	PostRepository postRepository;
-	
+
+	@Autowired
+	UserRepository userRepository;
+
 	@GetMapping
 	public List<CommentDto> getAllComments() {
 		List<Comment> comments = commentRepository.findAll();
 		return CommentDto.convert(comments);
 	}
-	
-	
-	
+
 	@PostMapping
-	public ResponseEntity<CommentDto> createComment(@RequestBody @Valid CommentsPostForm values, UriComponentsBuilder uriBuilder) {
-		System.out.println(values);
+	public ResponseEntity<CommentDto> createComment(@RequestBody @Valid CommentsPostForm values,
+			UriComponentsBuilder uriBuilder) {
 		Comment comments = values.convert(commentRepository, postRepository);
+
+		String username = SecurityContextHolder.getContext().getAuthentication().getName();
+		Optional<User> user = userRepository.findByUsername(username);
+		comments.setUser(user.get());
+
 		commentRepository.save(comments);
-		
+
 		URI uri = uriBuilder.path("/comments/{id}").buildAndExpand(comments.getId()).toUri();
 		return ResponseEntity.created(uri).body(new CommentDto(comments));
 	}
-	
+
 	@DeleteMapping("/{id}")
 	@Transactional
 	public ResponseEntity<?> remove(@PathVariable Long id) {
 		Optional<Comment> optional = commentRepository.findById(id);
+
+		System.out.println("id do comment: " + id);
+
 		if (optional.isPresent()) {
-			commentRepository.deleteById(id);
-			return ResponseEntity.ok().build();
+
+			// get the current user
+			String username = SecurityContextHolder.getContext().getAuthentication().getName();
+			Optional<User> user = userRepository.findByUsername(username);
+			Long currentUser = user.get().getId();
+
+			// Checking the owner of the comment
+			Long owner = commentRepository.findByUser();
+
+			if (currentUser == owner) {
+				commentRepository.deleteById(id);
+				return ResponseEntity.ok().build();
+			}
+
 		}
-		
 		return ResponseEntity.notFound().build();
 	}
